requirements.txt
mss
openai
python-dotenv
requests